+++ src/local/BroadcastMessage.d.ts
@@ -1,0 +1,4 @@
+export interface BroadcastMessage {
+    type: 'error' | 'update' | 'subscribed';
+    result: any;
+}