+++ src/local/Subscriber.d.ts
@@ -1,0 +1,5 @@
+import { BroadcastMessage } from "./BroadcastMessage";
+export interface Subscriber {
+    broadcastMessage(message: BroadcastMessage): void;
+    getConfig(): [];
+}